Final-Submission-Wu-Tang-Game
Unity Version 4.5.3f3
Chase Starks 5/12/2015
http://www.digitaltutors.com/software/Unity-tutorials

Setting up this project took a lot of my time and effort. The game starts off on the main menu where there are three levels that you can choose from. You will be able to grab hearts for health in case you are low on health and also collect fruits for points.

The Main features in the game are fruit collecting & hearts to collect. It come with simple moving left, right , up , & down. you must dodge enemies as they jump from the water to try to hurt yo. Some levels do have spikes where you must dodge them or climb around them to get the fruit. The project is loosely based on Mario and how simple it is because all you do is move side to side.